requirements.txt
pillow
numpy
# optional: onnxruntime==1.17.3
fastapi>=0.110.0
uvicorn>=0.29.0
onnxruntime>=1.17.0
reportlab>=4.2.0
